# api/docs/openapi.json
{
  "components": {
    "schemas": {
      "ApiMeta": {
        "properties": {
          "page": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "pageSize": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "requestId": {
            "type": "string"
          },
          "total": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "requestId"
        ],
        "type": "object"
      },
...
      "CreateStatusRequest": {
        "properties": {
          "color": {
            "type": "string"
          },
          "isDone": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "isHidden": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "color"
        ],
        "type": "object"
      },
      "CreateTagRequest": {
        "properties": {
          "color": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "color"
        ],
        "type": "object"
      },
...
      "ReorderStatusesRequest": {
        "properties": {
          "orderedStatusIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "orderedStatusIds"
        ],
        "type": "object"
      },
...
      "StatusDto": {
        "properties": {
          "color": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isDone": {
            "type": "boolean"
          },
          "isHidden": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int64",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "color",
          "sortOrder",
          "isDone",
          "isHidden",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "StatusListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/StatusDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "StatusResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/StatusDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "TagDto": {
        "properties": {
          "color": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "color",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "TagListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TagDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "TagResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TagDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "TaskDto": {
        "properties": {
          "archivedAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "dueDate": {
            "format": "date",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "position": {
            "format": "int64",
            "type": "integer"
          },
          "priority": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TaskPriority"
              }
            ]
          },
          "projectId": {
            "type": "string"
          },
          "startDate": {
            "format": "date",
            "type": "string"
          },
          "statusId": {
            "type": "string"
          },
          "tagIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "statusId",
          "title",
          "position",
          "startDate",
          "dueDate",
          "completedAt",
          "archivedAt",
          "tagIds",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "TaskPriority": {
        "enum": [
          "low",
          "medium",
          "high",
          "urgent"
        ],
        "type": "string"
      },
      "TaskResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TaskDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "UpdateStatusRequest": {
        "properties": {
          "color": {
            "type": [
              "string",
              "null"
            ]
          },
          "isDone": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "isHidden": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "UpdateTagRequest": {
        "properties": {
          "color": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local kanban task management API",
    "license": {
      "name": ""
    },
    "title": "Plan API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{projectId}/statuses": {
      "get": {
        "operationId": "listProjectStatuses",
        "parameters": [
          {
            "description": "Project id",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusListResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Statuses"
        ]
      },
      "post": {
        "operationId": "createProjectStatus",
        "parameters": [
          {
            "description": "Project id",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Statuses"
        ]
      }
    },
    "/api/v1/projects/{projectId}/statuses/reorder": {
      "post": {
        "operationId": "reorderProjectStatuses",
        "parameters": [
          {
            "description": "Project id",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderStatusesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusListResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Statuses"
        ]
      }
    },
    "/api/v1/projects/{projectId}/tags": {
      "get": {
        "operationId": "listProjectTags",
        "parameters": [
          {
            "description": "Project id",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagListResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ]
      },
      "post": {
        "operationId": "createProjectTag",
        "parameters": [
          {
            "description": "Project id",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ]
      }
    },
...
    "/api/v1/statuses/{statusId}": {
      "delete": {
        "operationId": "deleteStatus",
        "parameters": [
          {
            "description": "Status id",
            "in": "path",
            "name": "statusId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Statuses"
        ]
      },
      "patch": {
        "operationId": "updateStatus",
        "parameters": [
          {
            "description": "Status id",
            "in": "path",
            "name": "statusId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Statuses"
        ]
      }
    },
    "/api/v1/tags/{tagId}": {
      "delete": {
        "operationId": "deleteTag",
        "parameters": [
          {
            "description": "Tag id",
            "in": "path",
            "name": "tagId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ]
      },
      "patch": {
        "operationId": "updateTag",
        "parameters": [
          {
            "description": "Tag id",
            "in": "path",
            "name": "tagId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ]
      }
    },
...
    "/api/v1/tasks/{taskId}/tags/{tagId}": {
      "delete": {
        "operationId": "detachTaskTag",
        "parameters": [
          {
            "description": "Task id",
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Tag id",
            "in": "path",
            "name": "tagId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ]
      },
      "put": {
        "operationId": "attachTaskTag",
        "parameters": [
          {
            "description": "Task id",
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Tag id",
            "in": "path",
            "name": "tagId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://localhost:3001/api/v1"
    }
  ],
...
}